askracha/backend/requirements.txt
google-genai
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-readers-web
flask
flask-cors
python-dotenv
requests
beautifulsoup4
lxml
qdrant-client>=1.7.0
pytest>=7.0.0
APScheduler>=3.10.0